+++ app/Http/Controllers/PriceController.php
@@ -1,0 +1,131 @@
+<?php
+
+/**
+ * ausbilder.org - the free course management and planning software.
+ * Copyright (C) 2020 Holger Schmermbeck & others (see the AUTHORS file).
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+namespace App\Http\Controllers;
+
+use App\Price;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
+class PriceController extends Controller
+{
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware('verified');
+    }
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        abort_unless(Auth::user()->isAbleTo('price.edit', session('company_id')), 403);
+
+        $prices = Price::where('company_id', session('company_id'));
+
+        return $prices->get();
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
+    public function create()
+    {
+        abort_unless(Auth::user()->isAbleTo('price.edit', session('company_id')), 403);
+
+        return view('price.create');
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        abort_unless(Auth::user()->isAbleTo('price.edit', session('company_id')), 403);
+
+        $this->validate($request, [
+            'title' => 'required|min:3|max:255',
+            'price' => 'required',
+        ]);
+
+        $price = Price::create([
+            'company_id' => session('company_id'),
+            'title' => $request->title,
+            'price' => str_replace(',', '.', $request->price),
+            'description' => $request->description,
+            'currency' => 'EUR',
+        ]);
+
+        return $request;
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        //
+    }
+}