+++ database/migrations/2020_03_21_094534_create_participants_table.php
@@ -23,14 +23,18 @@ class CreateParticipantsTable extends Migration
             $table->string('street')->nullable();
             $table->string('zipcode', 5)->nullable();
             $table->string('location')->nullable();
+            $table->string('phone')->nullable();
             $table->string('email')->nullable();
             $table->dateTime('email_reminder')->nullable();
             $table->string('payee')->nullable();
             $table->boolean('participated')->default(0);
+            $table->decimal('price', 8, 2)->default(0);
+            $table->unsignedBigInteger('price_id');
             $table->boolean('payed')->default(0);
             $table->timestamps();
 
             $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
+            $table->foreign('price_id')->references('id')->on('prices');
         });
     }