+++ resources/views/emails/trainer-invite.blade.php
@@ -4,7 +4,7 @@
 {{ __('you got an invitation on ausbilder.org from :company.', ['company' => $company->name]) }}
 
 @component('mail::button', ['url' => env('APP_URL') . '/company/' . $company->hashid() . '/invite/' . $code])
-{{ __('view invitation') }}
+    {{ __('view invitation') }}
 @endcomponent
 
 {{ __("if you haven't an ausbilder.org account yet, register on for free.") }}